src/cloud_function/cf_products_inventory/requirements.txt
Faker==37.1.0
google-cloud-secret-manager==2.23.3
google-cloud-bigquery==3.35.0
functions_framework==3.8.3